backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.0
requests==2.31.0
redis==5.0.1
rq==1.15.1
gunicorn==21.2.0
pytest==7.4.2
black==23.9.1